+++ CHANGELOG.md
@@ -1,4 +1,13 @@
+## November 17, 2023
+
+### Backend
+
+* Remove code for Malaria ([@kystanleylin](https://github.com/kystanleylin))
+* Add email format checking ([@kystanleylin](https://github.com/kystanleylin))
+
 ## November 16, 2023
+
+### Backend
 
 * Split DbQuery into Malaria and SiteMgmt repos ([@kystanleylin](https://github.com/kystanleylin))
 * Add PUT and DELETE to Feedback and Action ([@kystanleylin](https://github.com/kystanleylin))
@@ -8,12 +17,16 @@
 
 ## November 15, 2023
 
+### Backend
+
 * Add Country table, ingested from REST Countries API ([@kystanleylin](https://github.com/kystanleylin))
 * Add WHO country metadata from DataPrep ([@kystanleylin](https://github.com/kystanleylin))
 * GET Malaria contains Country info ([@kystanleylin](https://github.com/kystanleylin))
 * Enable multi-args (separated by comma) in malaria/filter ([@kystanleylin](https://github.com/kystanleylin))
 
 ## November 11, 2023
+
+### Backend
 
 * Add malaria/filter and pagination ([@Mzisbrod](https://github.com/Mzisbrod))
 * Add error checking for 400 Bad Request ([@kystanleylin](https://github.com/kystanleylin))
@@ -22,5 +35,7 @@
 * Add admin/get_action ([@kystanleylin](https://github.com/kystanleylin))
 
 ## November 7, 2023
+
+### Backend
 
 * Add initial set of APIs and databases ([@kystanleylin](https://github.com/kystanleylin))